Add class Player to implement player controls. Player characters should inherit this.
